+++ src/IB.CSharpApiClient/Messages/TickByTickAllLastMessage.cs
@@ -8,12 +8,12 @@ namespace IB.CSharpApiClient.Messages
         public int TickType { get; private set; }
         public long Time { get; private set; }
         public double Price { get; private set; }
-        public long Size { get; private set; }
+        public decimal Size { get; private set; }
         public TickAttribLast TickAttribLast { get; private set; }
         public string Exchange { get; private set; }
         public string SpecialConditions { get; private set; }
 
-        public TickByTickAllLastMessage(int reqId, int tickType, long time, double price, long size, TickAttribLast tickAttribLast, string exchange, string specialConditions)
+        public TickByTickAllLastMessage(int reqId, int tickType, long time, double price, decimal size, TickAttribLast tickAttribLast, string exchange, string specialConditions)
         {
             ReqId = reqId;
             TickType = tickType;
