+++ src/IB.CSharpApiClient/Messages/UpdatePortfolioMessage.cs
@@ -5,7 +5,7 @@ namespace IB.CSharpApiClient.Messages
     public class UpdatePortfolioMessage
     {
         public Contract Contract { get; private set; }
-        public double Position { get; private set; }
+        public decimal Position { get; private set; }
         public double MarketPrice { get; private set; }
         public double MarketValue { get; private set; }
         public double AverageCost { get; private set; }
@@ -13,7 +13,7 @@ namespace IB.CSharpApiClient.Messages
         public double RealizedPNL { get; private set; }
         public string AccountName { get; private set; }
 
-        public UpdatePortfolioMessage(Contract contract, double position, double marketPrice, double marketValue, double averageCost, double unrealizedPNL, double realizedPNL, string accountName)
+        public UpdatePortfolioMessage(Contract contract, decimal position, double marketPrice, double marketValue, double averageCost, double unrealizedPNL, double realizedPNL, string accountName)
         {
             Contract = contract;
             Position = position;
